+++ CHANGELOG.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Outbound Identifiers / Outbound Identifier Lists
 - Roles (permissions)
 
+## [0.42.3] - 2019-01-11
+### Fixed
+- Invite now toggle on create user now sends the api call to set users status to invited
+### Added
+- Confirmation dialog for invite, resend invite, cancel invite,  and reset password
+
 ## [0.42.2] - 2019-01-11
 ### Fixed
 - Platform status on users page side panel now reflects the users invitation status